+++ tests/clients/test_broadcasts.py
@@ -1,7 +1,7 @@
 import pytest
 
 from berserk import Client
-from berserk.types.broadcast import BroadcastTop
+from berserk.types import BroadcastTop, PaginatedBroadcasts, BroadcastsByUser
 from utils import skip_if_older_3_dot_10, validate
 
 
@@ -11,3 +11,15 @@ class TestBroadcasts:
     def test_get_top(self):
         res = Client().broadcasts.get_top(page=1, html=False)
         validate(BroadcastTop, res)
+
+    @skip_if_older_3_dot_10
+    @pytest.mark.vcr
+    def test_search(self):
+        res = Client().broadcasts.search(query="chess", page=1)
+        validate(PaginatedBroadcasts, res)
+
+    @skip_if_older_3_dot_10
+    @pytest.mark.vcr
+    def test_get_by_user(self):
+        res = Client().broadcasts.get_by_user(username="lichess", page=1)
+        validate(BroadcastsByUser, res)