+++ CHANGELOG.rst
@@ -9,10 +9,15 @@ To be released
 
 * Added ``pgn_in_json`` parameter to ``client.games.export``.
 * Implement `broadcasts.get_top()` endpoint; typing fixes and validation.
+* Added ``client.broadcasts.search`` to search for broadcasts.
+* Added ``client.broadcasts.get_by_user`` to get broadcasts created by a user.
 * Added ``client.relations.block`` and ``client.relations.unblock`` for blocking/unblocking users.
 * Implemented ``/api/games/export/imports`` under
   ``client.games.export_imported``.
 * Added ``client.account.get_timeline`` to get current logged user timeline.
+* Implemented ``/api/study/by/{username}`` under
+  ``client.studies.get_by_user`` to get the metadata of all studies
+  of a given user.
 
 Thanks to all the contributors who helped to this release:
 - @hsheth2